Replace old automation page with automation1
